+++ src/main/java/me/villagerunknown/coinbanks/block/CoinBankBlocks.java
@@ -23,12 +23,53 @@ public class CoinBankBlocks {
 	public static final Block RED_TERRACOTTA_COIN_BANK;
 	public static final Block BLACK_TERRACOTTA_COIN_BANK;
 	
-	public static final Block COLD_PIGGY_BANK_BLOCK;
-	public static final Block TEMPERATE_PIGGY_BANK_BLOCK;
-	public static final Block WARM_PIGGY_BANK_BLOCK;
+	public static final Block COPPER_COIN_BANK;
+	public static final Block IRON_COIN_BANK;
+	public static final Block GOLD_COIN_BANK;
+	public static final Block DIAMOND_COIN_BANK;
+	public static final Block EMERALD_COIN_BANK;
+	public static final Block NETHERITE_COIN_BANK;
 	
-	public static final Block PIGLIN_BANK_BLOCK;
-	public static final Block ZOMBIFIED_PIGLIN_BANK_BLOCK;
+	public static final Block AMETHYST_COIN_BANK;
+	
+	public static final Block OBSIDIAN_COIN_BANK;
+	public static final Block CRYING_OBSIDIAN_COIN_BANK;
+	
+	public static final Block COLD_PIG_COIN_BANK;
+	public static final Block TEMPERATE_PIG_COIN_BANK;
+	public static final Block WARM_PIG_COIN_BANK;
+	
+	public static final Block PIGLIN_COIN_BANK;
+	public static final Block ZOMBIFIED_PIGLIN_COIN_BANK;
+	
+	public static final Block WHITE_SHEEP_COIN_BANK;
+	public static final Block ORANGE_SHEEP_COIN_BANK;
+	public static final Block MAGENTA_SHEEP_COIN_BANK;
+	public static final Block LIGHT_BLUE_SHEEP_COIN_BANK;
+	public static final Block YELLOW_SHEEP_COIN_BANK;
+	public static final Block LIME_SHEEP_COIN_BANK;
+	public static final Block PINK_SHEEP_COIN_BANK;
+	public static final Block GRAY_SHEEP_COIN_BANK;
+	public static final Block LIGHT_GRAY_SHEEP_COIN_BANK;
+	public static final Block CYAN_SHEEP_COIN_BANK;
+	public static final Block PURPLE_SHEEP_COIN_BANK;
+	public static final Block BLUE_SHEEP_COIN_BANK;
+	public static final Block BROWN_SHEEP_COIN_BANK;
+	public static final Block GREEN_SHEEP_COIN_BANK;
+	public static final Block RED_SHEEP_COIN_BANK;
+	public static final Block BLACK_SHEEP_COIN_BANK;
+	
+	public static final Block COW_COIN_BANK;
+	public static final Block RED_MOOSHROOM_COIN_BANK;
+	public static final Block BROWN_MOOSHROOM_COIN_BANK;
+	
+	public static final Block BEE_COIN_BANK;
+	public static final Block ANGRY_BEE_COIN_BANK;
+	
+	public static final Block RAVAGER_COIN_BANK;
+	
+	public static final Block GUARDIAN_COIN_BANK;
+	public static final Block ELDER_GUARDIAN_COIN_BANK;
 	
 	public CoinBankBlocks(){}
 	
@@ -51,12 +92,53 @@ public class CoinBankBlocks {
 		RED_TERRACOTTA_COIN_BANK = CoinBankBlocksFeatureLoader.registerJarCoinBankBlock( "red_terracotta_coin_bank" );
 		BLACK_TERRACOTTA_COIN_BANK = CoinBankBlocksFeatureLoader.registerJarCoinBankBlock( "black_terracotta_coin_bank" );
 		
-		COLD_PIGGY_BANK_BLOCK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "cold_piggy_bank" );
-		TEMPERATE_PIGGY_BANK_BLOCK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "temperate_piggy_bank" );
-		WARM_PIGGY_BANK_BLOCK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "warm_piggy_bank" );
+		COPPER_COIN_BANK = CoinBankBlocksFeatureLoader.registerJarCoinBankBlock( "copper_coin_bank" );
+		IRON_COIN_BANK = CoinBankBlocksFeatureLoader.registerJarCoinBankBlock( "iron_coin_bank" );
+		GOLD_COIN_BANK = CoinBankBlocksFeatureLoader.registerJarCoinBankBlock( "gold_coin_bank" );
+		DIAMOND_COIN_BANK = CoinBankBlocksFeatureLoader.registerJarCoinBankBlock( "diamond_coin_bank" );
+		EMERALD_COIN_BANK = CoinBankBlocksFeatureLoader.registerJarCoinBankBlock( "emerald_coin_bank" );
+		NETHERITE_COIN_BANK = CoinBankBlocksFeatureLoader.registerToughFireproofJarCoinBankBlock( "netherite_coin_bank" );
 		
-		PIGLIN_BANK_BLOCK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "piglin_bank" );
-		ZOMBIFIED_PIGLIN_BANK_BLOCK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "zombified_piglin_bank" );
+		AMETHYST_COIN_BANK = CoinBankBlocksFeatureLoader.registerJarCoinBankBlock( "amethyst_coin_bank" );
+		
+		OBSIDIAN_COIN_BANK = CoinBankBlocksFeatureLoader.registerToughJarCoinBankBlock( "obsidian_coin_bank" );
+		CRYING_OBSIDIAN_COIN_BANK = CoinBankBlocksFeatureLoader.registerToughJarCoinBankBlock( "crying_obsidian_coin_bank" );
+		
+		COLD_PIG_COIN_BANK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "cold_piggy_bank" );
+		TEMPERATE_PIG_COIN_BANK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "temperate_piggy_bank" );
+		WARM_PIG_COIN_BANK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "warm_piggy_bank" );
+		
+		PIGLIN_COIN_BANK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "piglin_bank" );
+		ZOMBIFIED_PIGLIN_COIN_BANK = CoinBankBlocksFeatureLoader.registerPiggyBankBlock( "zombified_piglin_bank" );
+		
+		WHITE_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "white_sheep_bank" );
+		ORANGE_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "orange_sheep_bank" );
+		MAGENTA_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "magenta_sheep_bank" );
+		LIGHT_BLUE_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "light_blue_sheep_bank" );
+		YELLOW_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "yellow_sheep_bank" );
+		LIME_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "lime_sheep_bank" );
+		PINK_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "pink_sheep_bank" );
+		GRAY_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "gray_sheep_bank" );
+		LIGHT_GRAY_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "light_gray_sheep_bank" );
+		CYAN_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "cyan_sheep_bank" );
+		PURPLE_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "purple_sheep_bank" );
+		BLUE_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "blue_sheep_bank" );
+		BROWN_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "brown_sheep_bank" );
+		GREEN_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "green_sheep_bank" );
+		RED_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "red_sheep_bank" );
+		BLACK_SHEEP_COIN_BANK = CoinBankBlocksFeatureLoader.registerSheepBankBlock( "black_sheep_bank" );
+		
+		COW_COIN_BANK = CoinBankBlocksFeatureLoader.registerCowBankBlock( "cow_bank" );
+		RED_MOOSHROOM_COIN_BANK = CoinBankBlocksFeatureLoader.registerCowBankBlock( "red_mooshroom_bank" );
+		BROWN_MOOSHROOM_COIN_BANK = CoinBankBlocksFeatureLoader.registerCowBankBlock( "brown_mooshroom_bank" );
+		
+		BEE_COIN_BANK = CoinBankBlocksFeatureLoader.registerBeeBankBlock( "bee_bank" );
+		ANGRY_BEE_COIN_BANK = CoinBankBlocksFeatureLoader.registerBeeBankBlock( "angry_bee_bank" );
+		
+		RAVAGER_COIN_BANK = CoinBankBlocksFeatureLoader.registerRavagerBankBlock( "ravager_bank" );
+		
+		GUARDIAN_COIN_BANK = CoinBankBlocksFeatureLoader.registerGuardianBankBlock( "guardian_bank" );
+		ELDER_GUARDIAN_COIN_BANK = CoinBankBlocksFeatureLoader.registerGuardianBankBlock( "elder_guardian_bank" );
 	}
 
 }